refactor: Add link to singleton API in Footer component
The Footer component in the src/components/Footer/Footer.tsx file has been updated to include a link to the singleton API. This change adds a new anchor tag with the href attribute set to the path generated by the makePath function for the "singleton" subdomain. This improves the navigation options in the Footer component.