refactor: Handle SVG file read error in Icon component
The Icon component in the src/components/Icon/Icon.tsx file has been updated to handle errors when reading SVG files. Previously, if there was an error reading the SVG file, it would throw an error and crash the application. Now, a try-catch block has been added to catch any errors and log them to the console. In case of an error, an empty string is assigned as the default value for the SVG content. This improves the robustness of the Icon component and prevents application crashes due to SVG file read errors.