refactor: Update Icon component to export IconTypes enum
The Icon component in the src/components/Icon/Icon.tsx file has been updated to export the IconTypes enum. This change allows other components or modules to import and use the IconTypes enum for type checking and specifying the type of icon to display. Exporting the enum improves the reusability and modularity of the Icon component.