refactor: Update MegaNav component to use data-driven approach
This commit refactors the MegaNav component to use a data-driven approach for rendering navigation items and sub-menus. Instead of hardcoding the items in the component, we now use an array of data objects to dynamically generate the navigation items and their corresponding sub-menus. This improves the maintainability and flexibility of the MegaNav component, allowing for easier updates and additions to the navigation menu.