refactor: Update ArticleItem component to make content prop optional
This commit updates the ArticleItem component in the ArticleItem.tsx file to make the content prop optional. Previously, the content prop was required, but now it can be omitted if necessary. This change provides more flexibility when using the ArticleItem component and avoids potential errors when the content is not available.