refactor: Update ArticlePage component to use dynamic heading level
This commit updates the ArticlePage.tsx file to use a dynamic heading level based on the index of each section. Previously, the heading level was hardcoded as either 1 or 2. Now, the level is determined by the index, with the first section having a level of 1 and subsequent sections having a level of 2. This change improves the flexibility and scalability of the ArticlePage component, allowing for easier addition and rearrangement of sections.